Refactor codebase to separate declarations and definitions
Moved function and class method implementations from header files to corresponding source files for improved code organization and compilation speed. Updated headers to contain only declarations, and added new .cpp files for Services, form, model, persistence, serializer, and related modules.
