+++ SKSE_Plugin/include/form_record.h
@@ -1,60 +1,21 @@
 #pragma once
-#include <cstdint>
 
 class FormRecord {
-    private:
      FormRecord() {}
-     FormRecord(RE::TESForm *_actualForm) {
-        actualForm = _actualForm;
-    }
-    public:
+     FormRecord(RE::TESForm *_actualForm);
 
-    static FormRecord *CreateNew(RE::TESForm *_actualForm, RE::FormType formType, RE::FormID formId) {
-        if (!_actualForm) {
-            return nullptr;
-        }
-        const auto result = new FormRecord(_actualForm);
-        result->formType = formType;
-        result->formId = formId;
-        return result;
-    }
-    static FormRecord *CreateReference(RE::TESForm *_actualForm) {
-        if (!_actualForm) {
-            return nullptr;
-        }
-        const auto result = new FormRecord(_actualForm);
-        result->formId = _actualForm->GetFormID();
-        return result;
-    }
-    static FormRecord *CreateDeleted(RE::FormID formId) {
-        const auto result = new FormRecord();
-        result->formId = formId;
-        result->deleted = true;
-        return result;
-    }
-    void UndeleteReference(RE::TESForm *_actualForm) {
-        if (!_actualForm) {
-            return;
-        }
-        deleted = false;
-        actualForm = _actualForm;
-        formId = _actualForm->GetFormID();
-    }
-    void Undelete(RE::TESForm *_actualForm, RE::FormType _formType) {
-        if (!_actualForm) {
-            return;
-        }
-        actualForm = _actualForm;
-        formType = _formType;
-        deleted = false;
-    }
+public:
 
-    bool Match(RE::TESForm *matchForm) {
-        if (!matchForm) {
-            return false;
-        }
-        return matchForm->GetFormID() == formId;
-    }
+    static FormRecord *CreateNew(RE::TESForm *_actualForm, RE::FormType formType, RE::FormID formId);
+    static FormRecord *CreateReference(RE::TESForm *_actualForm);
+
+    static FormRecord *CreateDeleted(RE::FormID formId);
+
+    void UndeleteReference(RE::TESForm *_actualForm);
+
+    void Undelete(RE::TESForm *_actualForm, RE::FormType _formType);
+
+    bool Match(RE::TESForm *matchForm);
 
     bool Match(RE::FormID matchForm) { return matchForm == formId; }
     bool HasModel() { return modelForm != nullptr; }
