+++ SKSE_Plugin/include/model.h
@@ -2,6 +2,7 @@
 #include <vector>
 #include <functional>
 #include <cstdint>
+#include "form_record.h"
 
 inline std::vector<FormRecord*> formData;
 inline std::vector<FormRecord*> formRef;
@@ -12,66 +13,20 @@ inline uint32_t firstFormId = 0;
 inline uint32_t dynamicModId = 0;
 
 
-inline void AddFormData(FormRecord* item) {
-    if (!item) {
-        return;
-    }
-    formData.push_back(item);
-}
+void AddFormData(FormRecord* item);
 
-inline void AddFormRef(FormRecord* item) {
-    if (!item) {
-        return;
-    }
-    formRef.push_back(item);
-}
+void AddFormRef(FormRecord* item);
 
-inline void EachFormData(std::function<bool(FormRecord*)> const& iteration) {
-    for (const auto item : formData) {
-        if (!iteration(item)) {
-            return;
-        }
-    }
-}
+void EachFormData(std::function<bool(FormRecord*)> const& iteration);
 
-inline void EachFormRef(std::function<bool(FormRecord*)> const& iteration) {
-    for (const auto item : formRef) {
-        if (!iteration(item)) {
-            return;
-        }
-    }
-}
+void EachFormRef(std::function<bool(FormRecord*)> const& iteration);
 
 
-inline void incrementLastFormID() {
-    ++lastFormId;
-}
+void incrementLastFormID();
 
 
-inline void UpdateId() {
-    EachFormData([&](FormRecord* item) {
-        if (item->formId > lastFormId) {
-            lastFormId = item->formId;
-        }
-        return true;
-    });
-    incrementLastFormID();
-}
+void UpdateId();
 
-inline void ResetId() {
-    lastFormId = firstFormId;
-}
+void ResetId();
 
-inline void ReadFirstFormIdFromESP() {
-    const auto dataHandler = RE::TESDataHandler::GetSingleton();
-
-    const auto id = dataHandler->LookupFormID(0x800, "Dynamic Persistent Forms.esp");
-
-    if (id != 0) {
-        espFound = true;
-    }
-
-    firstFormId = id + 1;
-    lastFormId = firstFormId;
-    dynamicModId = (firstFormId >> 24) & 0xff;
-}
+void ReadFirstFormIdFromESP();
