+++ SKSE_Plugin/src/form_record_serializer.cpp
@@ -1,0 +1,1 @@
+#include "form_record_serializer.h"
