+++ SKSE_Plugin/src/form_serializer.cpp
@@ -1,0 +1,1 @@
+#include "form_serializer.h"
